requirements.txt
aniso8601==7.0.0
anyio==3.6.1
asgiref==3.5.2
atomicwrites==1.4.0
attrs==21.4.0
autopep8==1.6.0
backpack==0.1
blinker==1.4
certifi==2022.6.15
charset-normalizer==2.0.12
cleo==0.6.8
click==7.1.2
colorama==0.4.4
Faker==0.8.18
fastapi==0.61.1
gmpy2==2.1.2
graphene-file-upload==1.3.0
graphene-pydantic==0.1.0
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.13.0
httptools==0.4.0
idna==3.3
inflection==0.3.1
iniconfig==1.1.1
lazy-object-proxy==1.7.1
mpmath==1.2.1
orator==0.9.9
packaging==21.3
pastel==0.1.1
pendulum==1.5.1
pluggy==0.13.1
promise==2.3
psycopg2-binary==2.8.6
py==1.11.0
pyaml==16.12.2
pycodestyle==2.8.0
pydantic==1.6
Pygments==2.12.0
pylev==1.4.0
pyparsing==3.0.9
pyserial==3.5
pytest==6.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-gql==0.2.4
pytz==2022.1
pytzdata==2020.1
PyYAML==5.4.1
requests==2.28.0
Rx==1.6.1
simplejson==3.17.6
six==1.16.0
sniffio==1.2.0
starlette==0.13.6
starlette-graphene3==0.5.1
starlette-graphql==0.2.1
sympy==1.10
text-unidecode==1.2
toml==0.10.2
typing_extensions==4.2.0
tzlocal==1.5.1
urllib3==1.26.9
uvicorn==0.12.2
watchgod==0.8.2
websockets==10.3
wrapt==1.14.1
